Create Screenshot (285).png
In our dashboard with #COVID-19 tweets, we are also dealing with the other tweets like government tweets as well as we considered some big cites of India also because this pandemic affect the country in all aspect.
It will analyze the government profile tweet every time and all those tweets will be analyzed based on polarity and emojis.
This project will be easily accessible as it will deploy on the cloud and from every device, it will be accessible.